+++ apps/laboratory/tests/shared/fixtures/w3m-wallet-fixture.ts
@@ -2,6 +2,9 @@ import { testM as base, testMSiwe as siwe } from './w3m-fixture'
 import { WalletPage } from '../pages/WalletPage'
 import { WalletValidator } from '../validators/WalletValidator'
 
+import { DEFAULT_SESSION_PARAMS } from '../constants'
+import { doActionAndWaitForNewPage } from '../utils/actions'
+
 // Declare the types of fixtures to use
 interface ModalWalletFixture {
   walletPage: WalletPage
@@ -9,11 +12,11 @@ interface ModalWalletFixture {
 }
 
 // MW -> test Modal + Wallet
-export const testMW = base.extend<ModalWalletFixture>({
-  walletPage: async ({ context }, use) => {
-    // Use a new page, to open alongside the modal
-    const walletPage = new WalletPage(await context.newPage())
-    await walletPage.load()
+export const testConnectedMW = base.extend<ModalWalletFixture>({
+  walletPage: async ({ context, modalPage }, use) => {
+    const page = await doActionAndWaitForNewPage(modalPage.clickWalletDeeplink(), context)
+    const walletPage = new WalletPage(page)
+    await walletPage.handleSessionProposal(DEFAULT_SESSION_PARAMS)
     await use(walletPage)
   },
   walletValidator: async ({ walletPage }, use) => {
@@ -23,7 +26,6 @@ export const testMW = base.extend<ModalWalletFixture>({
 })
 export const testMWSiwe = siwe.extend<ModalWalletFixture>({
   walletPage: async ({ context }, use) => {
-    // Use a new page, to open alongside the modal
     const walletPage = new WalletPage(await context.newPage())
     await walletPage.load()
     await use(walletPage)
@@ -33,4 +35,5 @@ export const testMWSiwe = siwe.extend<ModalWalletFixture>({
     await use(walletValidator)
   }
 })
+
 export { expect } from '@playwright/test'