+++ packages/ui/src/utils/TypeUtil.ts
@@ -188,7 +188,7 @@ export type LogoType =
 
 export type PlacementType = 'bottom' | 'left' | 'right' | 'top'
 
-export type ChipType = 'fill' | 'shade' | 'shadeSmall' | 'transparent' | 'success'
+export type ChipType = 'fill' | 'shade' | 'shadeSmall' | 'transparent' | 'success' | 'error'
 
 export type ButtonType = 'accent' | 'fill' | 'shade' | 'fullWidth' | 'accentBg'