+++ packages/scaffold-vue/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @web3modal/scaffold-vue
 
+## 4.0.11
+
+### Patch Changes
+
+- Analytics connection event improvements. Unsupported chain flag. Siwe package refactor. RPC improvements. UI improvements'
+
+- Updated dependencies []:
+  - @web3modal/scaffold@4.0.11
+
+## 4.0.10
+
+### Patch Changes
+
+- Add error state to wui-chip composite
+
+- Updated dependencies []:
+  - @web3modal/scaffold@4.0.10
+
 ## 4.0.9
 
 ### Patch Changes